+++ src/MineCase.Server.Interfaces/World/IWorld.cs
@@ -29,41 +29,4 @@ namespace MineCase.Server.World
 
         Task<GeneratorSettings> GetGeneratorSettings();
     }
-
-    public static class WorldExtensions
-    {
-        public static Task<BlockState> GetBlockState(this IWorld world, IGrainFactory grainFactory, int x, int y, int z)
-        {
-            var xOffset = MakeRelativeBlockOffset(x);
-            var zOffset = MakeRelativeBlockOffset(z);
-            var chunkColumnKey = world.MakeChunkColumnKey(xOffset.chunk, zOffset.chunk);
-            return grainFactory.GetGrain<IChunkColumn>(chunkColumnKey).GetBlockState(xOffset.block, y, zOffset.block);
-        }
-
-        public static Task SetBlockState(this IWorld world, IGrainFactory grainFactory, int x, int y, int z, BlockState blockState)
-        {
-            var xOffset = MakeRelativeBlockOffset(x);
-            var zOffset = MakeRelativeBlockOffset(z);
-            var chunkColumnKey = world.MakeChunkColumnKey(xOffset.chunk, zOffset.chunk);
-            return grainFactory.GetGrain<IChunkColumn>(chunkColumnKey).SetBlockState(xOffset.block, y, zOffset.block, blockState);
-        }
-
-        public static (int chunkX, int chunkZ) GetChunk(this Position blockPosition)
-        {
-            return (MakeRelativeBlockOffset(blockPosition.X).chunk, MakeRelativeBlockOffset(blockPosition.Z).chunk);
-        }
-
-        private static (int chunk, int block) MakeRelativeBlockOffset(int value)
-        {
-            var chunk = value / ChunkConstants.BlockEdgeWidthInSection;
-            var block = value % ChunkConstants.BlockEdgeWidthInSection;
-            if (block < 0)
-            {
-                chunk--;
-                block += ChunkConstants.BlockEdgeWidthInSection;
-            }
-
-            return (chunk, block);
-        }
-    }
 }
